New GUI three new buttons that return the desired table
search.html has a violation error as it searches first for all unique
names for any documnet
in the collection.
mongodb atlas needs to query all documents so that my program
can list them in a efficient way.
can Mongo handle this? it might be better to search for unique
positional number and return thier names.